add decimal to cy, cx and r

+++ db/migrate/20210926125615_create_data_visualization_items.rb
@@ -1,9 +1,9 @@
 class CreateDataVisualizationItems < ActiveRecord::Migration[6.1]
   def change
     create_table :data_visualization_items do |t|
-      t.integer :cx
-      t.integer :cy
-      t.integer :r
+      t.decimal :cx
+      t.decimal :cy
+      t.decimal :r
       t.string :data_id
       t.string :cluster
       t.references :data_visualization, null: false, foreign_key: true
